requirements.txt
Flask-SQLAlchemy==2.5.1
SQLAlchemy~=1.4.36
PyYAML~=5.4.1
Flask~=2.0.3
pymysql==1.0.2
pandas~=1.1.5
boto3~=1.22.7
botocore~=1.25.7
s3path~=0.3.4
scipy~=1.5.4
scikit-learn~=0.24.2
numpy~=1.19.5
joblib~=1.1.0
pytest~=7.0.1